Add analysis code and Labview code documentation
I changed the figure numbers for the data to be the final figure # on the submitted paper.  For the DyMIN_RT_Main code and the sub-vi's called by it I added a back and front panel pdf with just the one view, then added full documentation so Github users can see all the options in diagram and fully interpret the code even without a Labview license.